feat: add get pres pos function

+++ src/actuator_driver/include/actuator_driver/actuator_manager.hpp
@@ -30,7 +30,7 @@ class ActuatorManager
 		uint8_t* createPacket(uint8_t id, uint8_t instr, uint8_t* parameters, uint8_t parameter_size, uint8_t& out_size);
 		int readPacket(uint8_t* packet);
 		int readStatus(uint8_t id, StatusPacket& out);
-		uint8_t getPresentPosition(uint8_t id);
+		uint16_t getPresentPosition(uint8_t id);
 		
 		std::shared_ptr<SerialHandler> serial_handler_;	
 };
